+++ lib/upipe-hbrmt/sdienc.c
@@ -24,10 +24,10 @@
 
 #include "sdienc.h"
 
-void upipe_uyvy_to_sdi_c(uint8_t *dst, const uint8_t *y, int64_t pixels)
+void upipe_uyvy_to_sdi_c(uint8_t *dst, const uint8_t *y, uintptr_t pixels)
 {
     struct ubits s;
-    int64_t size = pixels * 2; /* change to number of samples */
+    uintptr_t size = pixels * 2; /* change to number of samples */
     ubits_init(&s, dst, size * 10 / 8, UBITS_WRITE);
 
     for (int i = 0; i < size; i ++)
